moved end of year calculations to separate module
